pyproject.toml
[tool.poetry]
name = "well-docs"
version = "0.1.0"
description = "well documantation build"
authors = ["SizNi <wwwwwwwww@list.ru>"]
readme = "README.md"
packages = [{include = "well_docs"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
gunicorn = "^20.1.0"
django-bootstrap5 = "^22.2"
python-dotenv = "^0.21.1"
dj-database-url = "^1.2.0"
django-tables2 = "^2.5.1"
pycairo = "^1.23.0"
ipython = "^8.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
